Preliminary refactoring into cmd and pkg
Move several functions to pkg/flakefinder. Also move the main program
folders to cmd/ subdirectory. Adjust make files and bash scripts.
