+++ robots/cmd/flake-report-writer/BUILD.bazel
@@ -3,18 +3,15 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "downloader.go",
         "index.go",
         "main.go",
         "report.go",
     ],
-    importpath = "kubevirt.io/project-infra/robots/flakefinder",
+    importpath = "kubevirt.io/project-infra/robots/cmd/flake-report-writer",
     visibility = ["//visibility:public"],
     deps = [
         "//robots/pkg/flakefinder:go_default_library",
         "@com_github_google_go_github_v28//github:go_default_library",
-        "@com_github_joshdk_go_junit//:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go_storage//:go_default_library",
         "@io_k8s_test_infra//prow/config/secret:go_default_library",
         "@io_k8s_test_infra//prow/flagutil:go_default_library",
@@ -42,7 +39,7 @@ container_push(
 )
 
 go_binary(
-    name = "flakefinder",
+    name = "flake-report-writer",
     embed = [":go_default_library"],
     visibility = ["//visibility:public"],
 )
@@ -50,16 +47,14 @@ go_binary(
 go_test(
     name = "go_default_test",
     srcs = [
-        "downloader_test.go",
-        "flakefinder_suite_test.go",
         "index_test.go",
+        "main_suite_test.go",
         "main_test.go",
         "report_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//robots/pkg/flakefinder:go_default_library",
-        "@com_github_google_go_github_v28//github:go_default_library",
         "@com_github_onsi_ginkgo//:go_default_library",
         "@com_github_onsi_ginkgo//extensions/table:go_default_library",
         "@com_github_onsi_gomega//:go_default_library",
