+++ robots/cmd/flake-report-writer/main.go
@@ -26,12 +26,10 @@ import (
 	"log"
 	"net/url"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"cloud.google.com/go/storage"
 	"github.com/google/go-github/v28/github"
-	"github.com/sirupsen/logrus"
 	"golang.org/x/oauth2"
 	"k8s.io/test-infra/prow/config/secret"
 	"k8s.io/test-infra/prow/flagutil"
@@ -120,120 +118,27 @@ func main() {
 	)
 	tc := oauth2.NewClient(ctx, ts)
 
-	c := github.NewClient(tc)
+	ghClient := github.NewClient(tc)
 
-	endOfReport := time.Now()
-	startOfReport, endOfReport := GetReportInterval(o, endOfReport)
-	logrus.Infof("Fetching Prs from %v till %v", startOfReport, endOfReport)
-
-	logrus.Infof("Filtering PRs for base branch %s", PRBaseBranch)
-	prs := []*github.PullRequest{}
-	prNumbers := []int{}
-	for nextPage := 1; nextPage > 0; {
-		pullRequests, response, err := c.PullRequests.List(ctx, o.org, o.repo, &github.PullRequestListOptions{
-			Base:        PRBaseBranch,
-			State:       "closed",
-			Sort:        "updated",
-			Direction:   "desc",
-			ListOptions: github.ListOptions{Page: nextPage},
-		})
-		if err != nil {
-			log.Fatalf("Failed to fetch PRs for page %d: %v.", nextPage, err)
-		}
-		nextPage = response.NextPage
-		for _, pr := range pullRequests {
-			if startOfReport.After(*pr.UpdatedAt) {
-				nextPage = 0
-				break
-			}
-			if pr.MergedAt == nil {
-				continue
-			}
-			if startOfReport.After(*pr.MergedAt) {
-				continue
-			}
-			if endOfReport.Before(*pr.MergedAt) {
-				continue
-			}
-			logrus.Infof("Adding PR %v '%v' (updated at %s)", *pr.Number, *pr.Title, pr.UpdatedAt.Format(time.RFC3339))
-			prs = append(prs, pr)
-			prNumbers = append(prNumbers, *pr.Number)
-		}
-	}
-	logrus.Infof("%d pull requests found.", len(prs))
-
-	client, err := storage.NewClient(ctx)
+	storageClient, err := storage.NewClient(ctx)
 	if err != nil {
 		log.Fatalf("Failed to create new storage client: %v.\n", err)
 	}
-	reports := []*Result{}
-	for _, pr := range prs {
-		r, err := FindUnitTestFiles(ctx, client, flakefinder.BucketName, strings.Join([]string{o.org, o.repo}, "/"), pr, startOfReport, o.skipBeforeStartOfReport)
-		if err != nil {
-			log.Printf("failed to load JUnit file for %v: %v", pr.Number, err)
-		}
-		reports = append(reports, r...)
-	}
 
-	err = WriteReportToBucket(ctx, client, reports, o.merged, o.org, o.repo, prNumbers, o.stdout, o.isDryRun, startOfReport, endOfReport)
+	reportBaseData := flakefinder.GetReportBaseData(ctx, ghClient, storageClient, flakefinder.NewReportBaseDataOptions(PRBaseBranch, o.today, o.merged, o.org, o.repo, o.skipBeforeStartOfReport))
+
+	err = WriteReportToBucket(ctx, storageClient, o.merged, o.org, o.repo, o.stdout, o.isDryRun, reportBaseData)
 	if err != nil {
 		log.Fatal(fmt.Errorf("failed to write report: %v", err))
 		return
 	}
 
 	printIndexPageToStdOut := o.isDryRun && o.stdout
-	err = CreateReportIndex(ctx, client, o.org, o.repo, printIndexPageToStdOut)
+	err = CreateReportIndex(ctx, storageClient, o.org, o.repo, printIndexPageToStdOut)
 	if err != nil {
 		log.Fatal(fmt.Errorf("failed to create report index page: %v", err))
 		return
 	}
-}
-
-func GetReportInterval(o options, till time.Time) (startOfReport, endOfReport time.Time) {
-	if o.today {
-		startOfReportToday := till.Format("2006-01-02") + "T00:00:00Z"
-		startOfReport, err := time.Parse(time.RFC3339, startOfReportToday)
-		if err != nil {
-			log.Fatalf("Failed to parse time %+v: %+v", startOfReportToday, err)
-		}
-		return startOfReport, till
-	} else {
-		startOfReport = till.Add(-o.merged)
-	}
-
-	// we normalize the start of the report against start of day vs. start of the hour to avoid working against a
-	// moving target.
-	// In general a user would expect to find all pull requests of the previous day in a 24h report, regardless of
-	// when the report has been run at the current day, which, depending on time of day when the report had been run,
-	// would not always be the case.
-	// Consider i.e. if the report is run late in the afternoon the user might wonder why the PR merged in the morning
-	// the day before was not included.
-
-	var startOfDayOrHour, endOfDayOrHour string
-
-	// in case of reports for at least a day we are fetching reports from start of previous day till end of that day
-	if o.merged.Hours() < 24 {
-		// in case of less than a day we are fetching reports from start of the hour
-		startOfDayOrHour = startOfReport.Format("2006-01-02T15:00:00Z07:00")
-		endOfDayOrHour = till.Format("2006-01-02T15:00:00Z07:00")
-	} else {
-		startOfDayOrHour = startOfReport.Format("2006-01-02") + "T00:00:00Z"
-		endOfDayOrHour = till.Format("2006-01-02") + "T00:00:00Z"
-	}
-	startOfReport, err := time.Parse(time.RFC3339, startOfDayOrHour)
-	if err != nil {
-		log.Fatalf("Failed to parse time %+v: %+v", startOfDayOrHour, err)
-	}
-	endOfReport, err = time.Parse(time.RFC3339, endOfDayOrHour)
-	if err != nil {
-		log.Fatalf("Failed to parse time %+v: %+v", endOfDayOrHour, err)
-	}
-	millisecond, err := time.ParseDuration("1ms")
-	if err != nil {
-		log.Fatalf("Failed to parse duration 1ms: %+v", err)
-	}
-	endOfReport = endOfReport.Add(-millisecond)
-	return startOfReport, endOfReport
 }
 
 // BuildReportOutputPath creates the path to which the report will get written, considering also if we are in
