+++ robots/cmd/flake-report-writer/main_suite_test.go
@@ -1,0 +1,27 @@
+package main_test
+
+import (
+	"flag"
+	"os"
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+type options struct {
+	printTestOutput bool
+}
+
+var testOptions = options{}
+
+func TestMain(m *testing.M) {
+	flag.BoolVar(&testOptions.printTestOutput, "print_test_output", false, "Whether test output should be printed via logger")
+	flag.Parse()
+	os.Exit(m.Run())
+}
+
+func TestFlakefinder(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Flakefinder Main Suite")
+}
