+++ robots/cmd/indexpagecreator/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "kubevirt.io/project-infra/robots/indexpagecreator",
+    importpath = "kubevirt.io/project-infra/robots/cmd/indexpagecreator",
     visibility = ["//visibility:public"],
     deps = [
         "//robots/pkg/flakefinder:go_default_library",
