+++ robots/cmd/indexpagecreator/update-jobs-with-latest-indexpagecreator-image.sh
