+++ robots/pkg/flakefinder/BUILD.bazel
@@ -1,13 +1,36 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["flakefinder.go"],
+    srcs = [
+        "downloader.go",
+        "flakefinder.go",
+        "report_data.go",
+    ],
     importpath = "kubevirt.io/project-infra/robots/pkg/flakefinder",
     visibility = ["//visibility:public"],
     deps = [
+        "@com_github_google_go_github_v28//github:go_default_library",
+        "@com_github_joshdk_go_junit//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_google_cloud_go_storage//:go_default_library",
         "@org_golang_google_api//iterator:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "downloader_test.go",
+        "flakefinder_suite_test.go",
+        "flakefinder_test.go",
+        "report_data_test.go",
+    ],
+    embed = [":go_default_library"],
+    deps = [
+        "@com_github_google_go_github_v28//github:go_default_library",
+        "@com_github_onsi_ginkgo//:go_default_library",
+        "@com_github_onsi_ginkgo//extensions/table:go_default_library",
+        "@com_github_onsi_gomega//:go_default_library",
     ],
 )
